Option to enable tracers with user-specified BCs
In this commit, the module tracer_example.F90 has been modified
to include 3 passive tracers (tracer1, tracer2, and tracer3) whose
initial and boudary conditions are specified by the user via
netCDF files. The surface values of the tracers are update every tracer
time step. Only locations where the BC's value are > 0 are updated. That
is, the surface value of tracers are unchaged in locations where the
BC value is = 0.